requirements.txt
# NOTE: This file contains ALL dependencies including Airflow.
# For control plane only (no Airflow), use: requirements-control-plane.txt
# For testing only (no Airflow), use: requirements-test.txt

# Airflow (requires SQLAlchemy <2.0)
apache-airflow==2.8.1
apache-airflow-providers-amazon==8.16.0
apache-airflow-providers-microsoft-azure==8.4.0
apache-airflow-providers-mongo==5.3.2
apache-airflow-providers-mysql==5.5.2

# FastAPI and web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
# NOTE: Airflow 2.8.1 requires SQLAlchemy <2.0, so using 1.4.x for compatibility
sqlalchemy==1.4.53
alembic==1.13.1
pymysql==1.1.0
cryptography==42.0.2

# AWS
boto3==1.34.41
botocore==1.34.41

# Azure
azure-storage-blob==12.19.0
azure-identity==1.15.0

# MongoDB
pymongo==4.6.1

# Kafka/CDC
kafka-python==2.0.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Scheduling
APScheduler==3.10.4
pytz==2024.1

# Utilities
python-dateutil==2.8.2
croniter==2.0.1
